+++ app/src/main/kotlin/ac/mdiq/podcini/ui/dialog/FeedSortDialog.kt
@@ -10,20 +10,15 @@ import ac.mdiq.podcini.storage.model.FeedSortOrder
 import ac.mdiq.podcini.storage.model.FeedSortOrder.Companion.getSortOrder
 import ac.mdiq.podcini.ui.fragment.SubscriptionsFragment.Companion.feedOrderBy
 import ac.mdiq.podcini.ui.fragment.SubscriptionsFragment.Companion.feedOrderDir
-import ac.mdiq.podcini.util.Logd
 import ac.mdiq.podcini.util.EventFlow
 import ac.mdiq.podcini.util.FlowEvent
-import android.app.Dialog
-import android.content.DialogInterface
+import ac.mdiq.podcini.util.Logd
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.WindowManager
 import android.widget.CompoundButton
-import android.widget.FrameLayout
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
 open class FeedSortDialog : BottomSheetDialogFragment() {
@@ -110,29 +105,10 @@ open class FeedSortDialog : BottomSheetDialogFragment() {
 
     protected open fun onSelectionChanged() {}
 
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        val dialog = super.onCreateDialog(savedInstanceState)
-        dialog.setOnShowListener { dialogInterface: DialogInterface ->
-            val bottomSheetDialog = dialogInterface as BottomSheetDialog
-            setupFullHeight(bottomSheetDialog)
-        }
-        return dialog
-    }
-
     override fun onDestroyView() {
         Logd(TAG, "onDestroyView")
         _binding = null
         super.onDestroyView()
-    }
-
-    private fun setupFullHeight(bottomSheetDialog: BottomSheetDialog) {
-        val bottomSheet = bottomSheetDialog.findViewById<FrameLayout>(com.leinardi.android.speeddial.R.id.design_bottom_sheet)
-        if (bottomSheet != null) {
-            val behavior = BottomSheetBehavior.from(bottomSheet)
-            val layoutParams = bottomSheet.layoutParams
-            bottomSheet.layoutParams = layoutParams
-            behavior.state = BottomSheetBehavior.STATE_EXPANDED
-        }
     }
 
     private fun setFeedOrder(selected: String, dir: Int) {